backend/requirements.txt

fastapi
uvicorn[standard]
openai
pydub
gTTS
aiofiles
python-multipart
requests
google-cloud-speech
